Tried to simplify the enumerator code base by removing an interface class for the old enumerator, and by removing the use of the enumerator factory. This last action systematically cleans up the code.
All relevant tests succeed.

The enumerator factory is now only used in some simulator tools,
and the next_state generator.

The motivation to do this is that I tried to understand why
so much time is spent in protecting/unprotecting ATerms and
constructing/destructing classes. In order to understand this,
it helps to have one prototype enumerator, that can of course
occur in several guises.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9103 21252fe5-60e8-0310-8ff0-c1ae1001594e