Put logging library in the namespace mcrl2::log
Note that if the namespace mcrl2::log is not in scope (but mcrl2 is), one
needs to write e.g. mCRL2log(log::debug) instead of mCRL2log(debug).
Reason for this change is to allow writing of logging routines like the foll

void some_logging_routine(log_level_t level, ...)
{
  if(mCRL2logEnabled(level))
  {
    mCRL2log(level) << ... << std::endl;
    ...
  }
}

These kinds of routines were not possible because of the macro magic that was
previously used in mCRL2logEnabled and mCRL2log.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9466 21252fe5-60e8-0310-8ff0-c1ae1001594e