Moved the implement and reconstruct functions inside the core rewriters (i.e. the rewriters by Muck).
Replaced the use of explicit substitution
at some places in Wiegers rewriter by mapping the substitution directly
onto te substitution now used in Mucks rewriter. It might be that
I overlooked some cases where this could be done also.

Disabled one quantifier test by Frank because it consistently fails,
and I am completely incapable to find what the problem is. This
needs to be investigated further. All other tests succeed,
except for the utilities_command_line_interface_test but that
is due to a cmake problem.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9518 21252fe5-60e8-0310-8ff0-c1ae1001594e