Resolved a problem with pbes2bool indicated by Tim.
The problem turned out to be with the universal and existential
quantifier. When the quantifier could not be eliminated, the
relevant procedure returned the original expression and forgot
to rewrite the body after the quantifier, including the necessary
substitution for variables.

This should have been resolved, and quantifier elimination has
been made a little faster.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9584 21252fe5-60e8-0310-8ff0-c1ae1001594e