Add mutable_indexed_substitution.
This substitution assumes the existence of a function f: variable -> size_t,
such that for each variable a unique index is produced. The substitution
is stored as a vector v, such that, for variable x, v[f(x)] determines the
right hand side of x in the substitution.

The substitution is not enabled in the rewriter yet, as it needs more testing, and possibly debugging. Furthermore, performance of this substitution needs to be evaluated.

Inspiration for this substitution was taken from the old substitution implementation by Muck van Weerdenburg in the rewriters.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9820 21252fe5-60e8-0310-8ff0-c1ae1001594e