Implemented an iterator based ATmakeAppl function. This function is NOT used for the corresponding atermpp function, awaiting discussion with Wieger on the exact functionality. Esp. the use of iterators with side effects in an invocation of aterm_appl in ltsmin2.h is subject of debate.
As a result the flag DO_AGGRESSIVE_GARBAGE_COLLECT is moved
to mcrl2/aterm/memory.h. The complete toolset has been tested
with this flag. This revealed one problem (caused by me....)
which has already been checked in.

Adapted the lineariser such that it will not warn anymore
about the loss of time in the translation of the conditional operator,
as agreed in the last tool meeting. 
Adapted the tool page to reflect this.

Some minor cleanups.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9888 21252fe5-60e8-0310-8ff0-c1ae1001594e