Implement proof of concept for a unified toolset documentation.
The directory doc/sphinx contains everything that is needed for generating the following:
 * Library user manual
 * Library reference pages

Currently, the LTS library documentation has been converted to this system as an example.

To generate the documentation, you need the following tools:
* doxygen
* sphinx (http://sphinx.pocoo.org/)
* python
* xsltproc

Sphinx is used as the main documentation tool, and uses reStructuredText as its input format. The doxygen documentation is converted to reStructuredText using xsltproc.

To generate the documentation:
% cd doc/sphinx
% python generate_documentation.py

This will generate a directory doc/sphinx/_build, that contains an index.html file which can be opened in your web browser.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@10073 21252fe5-60e8-0310-8ff0-c1ae1001594e