Add extra argument type to the command line interface classes.
The type enum_argument allows for the enumeration of arguments, with descriptions.

An example is the usage for the rewriter type, where the options are:
void add_options(interface_description& desc)
{
  Tool::add_options(desc);

  desc.add_option("rewriter", make_enum_argument<data::rewriter::strategy>("NAME")
                  .add_value("jitty", "for jitty rewriting", true)
                  .add_value("jittyc", "for compiled jitty rewriting")
                  .add_value("jittyp", "for jitty rewriting with prover"),
                  "use rewrite strategy NAME:",'r');
}

This assumes there is an operator >> to data::rewriter::strategy, and jitty is the
default value.

This enables generation of improved XML, which will improve the documentation,
and can eventually replace the custom XML that is used by mcrl2-gui.

Some tools have already been adapted, but the majority still needs to be changed.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@10376 21252fe5-60e8-0310-8ff0-c1ae1001594e