Modify logger to model-view pattern
This change was requested by Sjoerd and Frank in order to build a decent widget for logging in the graphical tools.

Multiple output policies can be registered with the logger (by default the file_output policy is registered). Of all registered output policies, the output(...) method is called if a log message must be processed.

For now I also implemented a function_pointer_output_policy, that substitutes the custom message handler that was implemented previously. This should be removed once the new functionality is used in the graphical tools.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@10400 21252fe5-60e8-0310-8ff0-c1ae1001594e