Drastically simplified garbage collection. In particular removed the usage of different generations of garbage. The main reason to carry this change out was that the implementation used lots of code duplication, did wast memory (because blocks of memory that were marked old could not be reused, except if they were completely emptied; code to promote old blocks to young was not actually used) and did not really seem to be effective (because most blocks were not marked old, garbage collection of young terms was approximately as expensive as garbaging all terms).
Check this in now, as changes are delicate. It needs to be experienced
to find out whether the current changes have a substantial effect on
time usage and memory. Expectation is that there will hardly be any
influence on the time used, but memory usage for large applications
may require some fine tuning.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@10420 21252fe5-60e8-0310-8ff0-c1ae1001594e