Replaced the types ATerm, ATermAppl, ATermList and ATermInt which were pointers to the structs _ATerm, _ATermAppl, _ATermList and _ATermInt by proper classes.
This has a substantial effect, as ATerm's cannot be used as pointers 
anymore. In order to compare an ATerm t with for instance a NULL pointer, 
as is heavily done in the toolset, it must be transformed to a pointer by
&*t. It has been suggested to add an operator bool() to an aterm to check
whether it is well defined, but this leads to ambiguities.

Furthermore, a number of smaller problems have been settled. Comparing
multi_actions via permutations was incorrect.

Several tests used =, where == had to be used.

All tests succeed.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@10456 21252fe5-60e8-0310-8ff0-c1ae1001594e