I left two residual errors, one due to me in the previous commit.
When reading a .aut file the initial state and the value 0 were 
mixed up in the translation table.

When writing a .aut file generated using bit hashing, the initial 
state was set to 0, instead of to the bit hashed valued. This
led to a transition system with one state too many, and unreachable
transitions.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11065 21252fe5-60e8-0310-8ff0-c1ae1001594e