Compared the use of MCRL2_SYSTEM_SPECIFIC_ALLOCA versus static and non static use of vector in the loop in the enumerator. Results are somewhat unexpected. I used state space generation of 1394-fin.
MCRL2_SYSTEM_SPECIFIC_ALLOCA:   36.2s
static vector                   37.4s
vector                          36.3s

As a vector does not overflow the stack I left the stack in place.

I did not adapt the ugly pointer argument in ApplyArray, because
this requires to rewrite the old ATerm library to contain iterators.
These are present and used in the new merged atermpp library.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11088 21252fe5-60e8-0310-8ff0-c1ae1001594e