Fix nextstate generator check on conditions.
The new implementation of the nextstate generator failed to check whether the solutions provided by the nextstate generator are exact. This means that the nextstate generator would enable transitions with non-false guards, instead of transitions for which the guard was true.

Now if the solution is not exact, we fail with an mcrl2::runtime_error.

Fixes #1073 of which #1075 was a duplicate.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11122 21252fe5-60e8-0310-8ff0-c1ae1001594e