Fixed #1085.
A list of action names was assigned to an identifier string, 
which was not accepted in debug mode (but would not have caused
problems in release mode).

As the string was only used to add a postfix to names of newly 
introduced variables, and the names of the multi-actions were
not at all used for this purpose, the erroneous assignment has been
removed.

All self tests appear to succeed, including the new one indicating
this problem.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11161 21252fe5-60e8-0310-8ff0-c1ae1001594e