Fixes #1095
The problem is a missing declaration (ATermAppl) that is not properly generated
in the compiling rewriter because a variable arity does not reflect the arity
of the main function at some point. As a solution the cast (ATermAppl) is
generated always at the appropriate point. 

This fix must be incorporated in the release.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11233 21252fe5-60e8-0310-8ff0-c1ae1001594e