Improve output of BES to CWI format.
This commit allows saving a BES in CWI format, even if the initial variable of the BES is not the first in the container of equations. If the equation for the initial variable is in the first block, we simply swap it with the first equation, otherwise we introduce an additional equation that directly points the this initial variable.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11635 21252fe5-60e8-0310-8ff0-c1ae1001594e