Removed libraries/core/include/mcrl2/core/typecheck.h.
Enabled the action rename typechecker to use the new
typechecker with in-class variables. This required to replace
some types "action" by aterm_appl, as soundness checks
do not allow to use an untyped action, which is of
the form ParamId(identifier_string, data_expression_list).

Removed a substantial amount of code that is not used
anymore. 

Adapted a number of files, using the namespace atermpp,
without declaring it using the "using" keyword,
which was possible as "using atermpp" was declared in
core/typecheck.h.

All tests succeed, at least on the mac.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11645 21252fe5-60e8-0310-8ff0-c1ae1001594e