Finished adapting the typecheckers. If I did not overlook situations, there are no gsMake/gsIs functions in the typechecker anymore, terms are not selected by indexing of their arguments, and also all occurences of set_argument have been removed. Removed the use of the core namespace from the data type checker. There are also no explicit uses of aterm, aterm_appl or aterm_list in the typecheckers.
Removed the set_argument function from the atermpp libraries, as it 
was not used anymore (in accordance with a decision of appr. a month
ago). 

All tests succeed on the mac.

The only issue that remains is that some terms that are not properly
typed, are not properly printed, causing assertions going of in debug
mode, and probably worse things in release mode.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11718 21252fe5-60e8-0310-8ff0-c1ae1001594e