Resolve bug #978. This bug regards linearisation of processes with variables that are bound in a sum, which also occurs in the assignment syntax. E.g.
 proc R(a,b : Bool) =
    rd(a) . R()
    + rd(b) . R() + sum b:Bool. wr(b) . R(a=b); 

In the last invocation of R the a must be assigned the bound
variable b, wheras to b the non bound variable b must be assigned.
This was spotted by Matthias Raffelsieper.

The problem has been resolved, by applying a renaming of 
bound variables at a relevant point. Furthermore, the mixed
use of assignments and data expression lists to pass arguments
to parameters in processes has been replaced by assignments
whereever possible.

A major rebuild of parts of the lineariser was required for this
purpose. All tests succeed. A linearisation test has been added,
but this test does not check proper translation. It only checks
that translation does not crash.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11768 21252fe5-60e8-0310-8ff0-c1ae1001594e