Reimplemented the storage of function symbols such that they do not move around in memory anymore, when resizing the array for function symbols. In the new implementation blocks of memory chunks are iteratively assigned.
The new implementation is enabled using the flag

#define FUNCTION_SYMBOL_AS_POINTER

in the file 

libraries/atermpp/include/mcrl2/atermpp/detail/function_symbol.h

By default it is off.

The performance of this new implementation is quite comparable to
the old.

There are a few changes in the old implementation that turned out
te be necessary.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11823 21252fe5-60e8-0310-8ff0-c1ae1001594e