Enforce accuracy setting in ltsview at least 1.
The reshape routines require their parameter N, encoding the accuracy of the shape,
to be at least 1. This was not enforced by the code, and it was not documented,
but it is apparent from inspection of the code. Added assertions that check
this precondition.

I modified the settings panel in the UI such that it enforces a value of at
least 1 (where it previously also allowed 0), such that we satisfy the
precondition above.

Fixes #1181

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11856 21252fe5-60e8-0310-8ff0-c1ae1001594e