Reordered the sequence default terms are being made. It used to be that first constructor constants and then constant mappings were used. If that would fail a constructor function applied to default terms were considered. But constant mappings generally do not have the required rewrite rules. This caused the state space generation of bke.mcrl2 to go astray. Now preference is given to a constructor applied to default terms, above a constant mapping.
Adapted a test case that failed after this change.

Removed the use of boost bind from the jittyc compiler.
I used the c++11 function notation for this, which was
accepted by the mac compiler.
As far as I can see the jittyc compiler now does not
depend on boost anymore. This resolves bug report #990.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11904 21252fe5-60e8-0310-8ff0-c1ae1001594e