Fix lts_lps2lts_test.
Addresses #1196

This bug was triggered only in the event that there was *o*
summand with the action that was marked as confluent tau action.
A condition on the set of prioritised summands was used in the
initialisation, whereas other code relied on the flag that
denotes whether confluence is used.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11945 21252fe5-60e8-0310-8ff0-c1ae1001594e