Addresses #755. The GUI applications now try to attach to the parent console. This does not yet give the expected behaviour in the CMD shell, as the console is not controlled by the GUI application, but at least it is showing you command-line feedback now. A solution would be to add .COM files for every executable in Windows which are run in console mode (and have preference over .EXE files in CMD), which simply call the .EXE.
The console is now freed before the GUI is shown, which means that any log output that is generated while the GUI is visible is not redirected to the console anymore. This could be changed by moving the code in pre_run() (in qt_tool.h) to ~qt_tool(). With the current setup, it is also an option to compile the Windows GUI applications as console applications (using /SUBSYSTEM:CONSOLE). Then the applications would behave like expected in console mode, but would briefly flash a console window when started directly from Windows.

There is no really nice solution to this problem; the .COM-solution seems most popular.

A quick work-around for now when you want the console to wait on the GUI application: rather than running "tool.exe parameters" in CMD, run "start /wait tool.exe parameters".



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12016 21252fe5-60e8-0310-8ff0-c1ae1001594e