Add test case for a problem in the lineariser.
The problem was observed by Jeroen Redegeld, and analysed by JFG. A fix has been proposed, and this is the test that goes with it.

This adds a new type of test to the LPS library, in which a specification can
be linearised using a number of strategies; for the result an LTS is generated,
which is compared w.r.t. bisimilarity to an expected state space.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12091 21252fe5-60e8-0310-8ff0-c1ae1001594e