Added FSet(S) and FBag(S). Adapted the typechecker as good as possible. All tests succeed, including all release tests.
Adapting the type checker was a nuisance. It is likely
that there are numerous potential problems with
the type checker. The only solution is to replace
the data term type checker completely, as the current
software bulldozer software. Whenever a problem occurred
the software is bulldozered such that the problem does
not occur anymore.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12181 21252fe5-60e8-0310-8ff0-c1ae1001594e