Disabled a check for lost terms, as this crashes on some platforms as the aterm_administration may be partly removed.
Replaced some aterm_casts by down_casts.

Used the function recursive_number_of_args instead of arity,
as the former yields the correct number of arguments and
the latter does not for terms with nested applications.
This solves a problem occurring in the jittyc rewriter,
causing some release tests to crash. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12335 21252fe5-60e8-0310-8ff0-c1ae1001594e