Resolves #1256
In this error report it is indicated that count({s1:1,s1:1})
cannot be rewritten. 

The problem is that the generation of built in data types are based on 
scripts that do not employ typing. There was a rule 

  @fbag_insert(d, p, @fbag_cons(d, q, b))  =  @fbag_cons(d, +(p,q), b);

in the file fbag.spec in which the symbol + occurred. This was supposed
to be +:Pos#Pos->Pos. However, it was taken to be the + <"union_"> : FBag(S) <"left"> # FBag(S) <"right"> -> FBag(S); Hence the statement 1+1<3 which could not be rewritten
was a badly typed expression. The problem has been resolved by replacing the rule 
in fbag.spec by the following, which is even slightly more efficient:

+ <"union_"> : FBag(S) <"left"> # FBag(S) <"right"> -> FBag(S);

All tests succeed.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12400 21252fe5-60e8-0310-8ff0-c1ae1001594e