Repaired a few problems occurring when running the tests with the MCRL2_TEST_COMPILERS flag.
Problems were that default functions for FSet an FBag were not
added to a data type (i.e. if, ==, !=, <=, <, >, >=).

Partially rewritten functions with more elements than 4 were not
correctly used conform the declaration, causing the compilation of
the rewriter to fail.

Added a missing creation of a rewriter in a test. 

All tests pass, except the linearisation test, which
runs into a timeout with the above flag switched on.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12491 21252fe5-60e8-0310-8ff0-c1ae1001594e