Repaired some printing problems.
Gave the difference operator on an FSet "left" and "right"
as functions to obtain arguments, instead of arg1 and arg2

Added a print routine for "reduce_fraction".

Explicitly declared some functions in print.h as binary operators.

Added some code to deal with printing fsets.
It appears to me that the code for printing fsets is not
complete.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12673 21252fe5-60e8-0310-8ff0-c1ae1001594e