Resolved a problem with an invalid memory access brought to light by Maikel Leemans. The problem is documented in ticket #1279.
Comment:

This problem was due to the fact that references to n-ary
application aterm function symbols were used, where this array was
a vector that was relocated in memory. The problem only came to light
now as due to the use of applications with transformers on the
arguments application(f, begin, end, transformer) the application function
symbol was already obtained from the vector, before the transformer
was applied to the arguments. While the transformer was applied, the
vector was relocated, and the reference to the application symbol was
rendered invalid.

This has been resolved by replacing the vector by a deque, which is
guaranteed
not to be relocated, when growing in size.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12680 21252fe5-60e8-0310-8ff0-c1ae1001594e