Cleaned the classic enumerator up, after merging it with and removing EnumeratorSolutionsStandard.
Removed the legacy rewriter. It is not needed anymore.

Stored the substitution in the enumerator via a pointer,
to prevent it from being copied. Used enumerator.end(),
instead of enumerator.iterator() because the former does
not construct a new iterator in the next_state_generator.

It appears that the enumerator became a little faster
due to these operations.

All release tests succeed.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12694 21252fe5-60e8-0310-8ff0-c1ae1001594e