Resolved bug #1281 indicating that higher order rewriting in the compiling rewriter did not function properly.
The compiling rewriter did not properly rewrite
higher order functions with more than two arguments. 
It simply forgot to build a match tree when there
were more arguments, forgetting the first arguments.
Matching was subsequently done properly, leading
to erroneously matching of arguments.

Beside that the array to determine which terms were
not in normal form did not always yield the correct results,
as not always the correct number of arguments of a function
were calculated. This has been dealt with by isolating the
functions accessing the "always rewrite" array, and adding
extra asserts to know that accesses are always in this array.

The rewriting_test.cpp has been extended with higher order
rewriting tests. Some of the tests in rewriting_test were
outcommented. This has been restored. Some other tests
were not functional. This has also been repaired.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12742 21252fe5-60e8-0310-8ff0-c1ae1001594e