Removed yet another ad_hoc pbes rewriter and replaced it by the pbes_one_point_rule and pbes_simplify_quantifiers rewriter.
Found a serious (copy and paste) bug in the pbes_simplify_quantifiers rewriter.
This has been repaired.

Found that the pbes rewriters introduced != on data terms. These inequalities
had been removed from the rewriters to limit the size of the c++ code of the 
compiling rewriter. Now they are added by default.

Found out that the order of variables in quantifiers had a huge
effect on the speed of enumeration. For the nodeadlockcheck on 1394-fin ordering
leads to an increase of the speed pbes2bool solves a pbes with a factor 2. 
Added an ad hoc routine to optimise the order of variables in quantifiers in a PBES. 

Tests succeed.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12774 21252fe5-60e8-0310-8ff0-c1ae1001594e