Made the indexed_set a template class, such that it can be used for other classes than just aterms. Removed some unused features from the indexed set.
An indexed set does not depend essentially on aterm features,
and could be moved out of the aterm class.

Investigation shows that an indexed set is still 25% more
memory efficient than an unordered_map. As an indexed set
is crucial for the memory footprint, this is a reason not to
replace an indexed set by an unordered map.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12853 21252fe5-60e8-0310-8ff0-c1ae1001594e