Adapted the a balanced tree, such that it will not use a stack container anymore within its iterator. It now uses a fixed array, which it can put on a stack. This fixed array is 20 positions big, which means that there is a maximum of 1.000.000 terms that one can store in a state. The largest number of state parameters that I have ever seen was 50.000 for an extreme example, which could and cannot be verified at all.
All tests succeed. A very minor performance gain is observed compared to the version
with a stack container. The reason for removing the stack is that Wieger reports that
on linux creating and removing this stack is responsible for 15% of the running time
in examples that do not rely heavily on the this stack.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12880 21252fe5-60e8-0310-8ff0-c1ae1001594e