Adapted pbes2bool to save memory. Added a flag --erase to erase generated boolean equations that are not being used and which have a non trivial right hand side. In 5x5 four in a row with the options -zd and -s2 (depth first search, and substitute true/false for variables whenever possible) this leads to 50% reduction in memory.
Restored the possibility in an indexed set
to erase elements. In order to simplify the code
this option was removed, but it is necessary to
implement the --erase option, as generated equations
must be removed, and the used memory had to be released.

Made pbes2bool and pbes2bes rewriter tools, instead
of pbes rewriter tools. The option to select a pbes_rewriter
was not used in pbes2bool and pbes2bes.

Adapted the pbes2bool and pbes2bes tests to include
the new --search and --erase options, and removed
tests with various rewriters.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12996 21252fe5-60e8-0310-8ff0-c1ae1001594e