[Could be for the release as some non existing flags in the  example files have been removed, but this commit only adresses  examples and does not contain essential code changes]
Removed the -t or --tree flag from example python files, as
these are not used anymore. Made a small change in four in a row,
and added some experimental results. The only boards that
cannot be verified by pbes2bool as it stands are 7x5, 6x6 and 
6x7 (full board), although I actually believe that the current
version of pbes2bool on the mammoet should just be able to do
the 6x6 and 7x5 boards. This would require the generation of
2Gbes boolean equation systems.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13015 21252fe5-60e8-0310-8ff0-c1ae1001594e