[Release, this is a nasty bug]
The problem was due to the the fact that the illegal transition would end
in the initial state. A trace was printed to the state following the illegal
state, to include the illegal action in the trace. However, if this is the
initial state the empty trace is printed.

This has been resolved by adding the option to print a trace to a state
followed by adding an extra single transition.

This resolves bug #1306.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13058 21252fe5-60e8-0310-8ff0-c1ae1001594e