Resolve ticket #1324.
The problem was that a variable bound in a sum could become a process parameter. This typically occurs in the process of the shape

proc X = sum r:Nat. a.b(r);

Clearly, r must be recalled between the a and the b action.

It is unclear why this problem was caused, most likely due to the removal of an alphaconversion which seemed to be superfluous.
An explicit extra alphaconversion has been added to rename the variable in the sum, to remedy this situation. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13514 21252fe5-60e8-0310-8ff0-c1ae1001594e