Resolved #1326.
In this problem ltsconvert --tau=... =ebranching-bisim did not
work properly. 

The reason for this problem turns out to be that
an action with label "tau" was indicated NOT to be 
an internal action after the function hide_actions was
applied to it. The branching bisimulation routines 
would consider it as an external action, and therefore
the reduction did not work properly.

Added an assertion to detect this problem.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13597 21252fe5-60e8-0310-8ff0-c1ae1001594e