Removed the SVC library. From now on .lts files are saved as an aterm. This turns out to require slightly less diskspace and is much faster. Moreover, .lts files can from now on be used for stdin and stdout.
This operation saves appr. 4000 lines of (quite
incomprehesible) code, both in the lts library 
(where a data specification, process parameters and
action declarations were put at the end of an SVC
in a rather dirty manner, and in the SVC library
which was typically encoded in a mid-90-ies C style.

This makes it possible to extend .lts files to
include probabilities and time. Probabilities are
now only possible in .aut files. Time is not stored at
all in labelled transition system formats.

All tests succeed.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13666 21252fe5-60e8-0310-8ff0-c1ae1001594e