The lts format can now store lts's with probabilities. The internal format is not compatible with previous formats.
Moreover, the .dot and .fsm format can internally deal with
probabilities, too. It is not possible to save .dot, and to
save and load .fsm files, because the disk format has not yet
been adapted.

It is possible using ltsconvert to translate lts's with probabilities
from all formats to all formats.

Most tools operating on lts's cannot yet deal with probabilities.
Doing equivalence reduction or visualisation of probabilistic lts's
is most likely doomed to fail.

All tests succeed.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13734 21252fe5-60e8-0310-8ff0-c1ae1001594e