Replaced the use of size_t as numbers to represent the enumerator and the denominator of fractions in .fsm and .aut file by the use of big_natural_numbers. Big_natural_numbers are arbitrarily sized natural numbers, which can grow as big as computer memory allows. They are implemented as vectors of size_t and ought to be reasonable efficient. The implemenation of mod and div in the big numbers is not yet optimal.
An extensive test for big_numbers has been added, which succeeds. Unfortunately, as it stands
this is not executed, as in the CMakeFile for the directory testing of utilities tests have
been disabled by the NOTESTS keyword. This should be removed, but removing it leads
to a number of failing tests, which needs to be investigated first.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13776 21252fe5-60e8-0310-8ff0-c1ae1001594e