Adapted ltsgraph to use classes instead of structures for its elementary data structures, protecting the variables in the structures from direct access. Tried to use constructors, instead of direct assignments, to fill these datastructures whenever possible.
Introduced a new class "NodeWithColor" to remove a small amount of code
duplication. This is not to be negative about the code. The code is remarkably
accessible and very clearly written!!!

Started to prepare to make ltsgraph suitable for probabilistic transition
systems by adding the field "is_probabilistic" to a NodeNode, representing
a state. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14178 21252fe5-60e8-0310-8ff0-c1ae1001594e