Simplified mcrl2/lts/detail/fixed_vector.h: using private inheritance, it can now be guaranteed that the iterators never are invalidated because the methods that would do so are not available to the user.
Improved comments in mcrl2/lts/detail/liblts_bisim_gjkw.h and in
liblts_bisim_gjkw.cpp: they are now in line with the submitted version of the
article, by using ``RfnB'' instead of ``RefB'' (the latter was too similar to
``RedB'') and ``non-bottom'' instead of ``nonbottom''.

Found a few instances where #include "..." should be #include <...> and vice
versa.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14236 21252fe5-60e8-0310-8ff0-c1ae1001594e