This resolves bug #1366.
The representation generator was designed to not yield too complex representatives. The complexity
was bound to three, to avoid a too extensive search for representative terms for a sort. 

Based on this bug report the code has been adapted such that it will always generate a representative if
it exists. The idea is that a recursive search is done for a term of a certain sort, but when searching for
a term of some sort it is not allowed to search recursively for a term of that sort. Such a search can in theory
be exponential, but I expect that this will rarely occur in practice. 

All tests succeed. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14277 21252fe5-60e8-0310-8ff0-c1ae1001594e